pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sentinel-cybersecurity-triage"
version = "0.1.0"
description = "AWS-native cybersecurity news triage and publishing system"
authors = [
    {name = "Sentinel Team", email = "team@sentinel.local"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Security",
    "Topic :: Internet :: WWW/HTTP :: Indexing/Search",
]

dependencies = [
    "boto3>=1.34.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "feedparser>=6.0.10",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "PyYAML>=6.0.1",
    "python-dateutil>=2.8.2",
    "jsonschema>=4.20.0",
    "structlog>=23.2.0",
    "aws-lambda-powertools>=2.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
    "moto>=4.2.0",
    "black>=23.12.0",
    "isort>=5.13.0",
    "flake8>=6.1.0",
    "mypy>=1.8.0",
]

ml = [
    "nltk>=3.8.1",
    "scikit-learn>=1.3.0",
    "fuzzywuzzy>=0.18.0",
    "python-Levenshtein>=0.20.0",
]

export = [
    "openpyxl>=3.1.0",
    "xlsxwriter>=3.1.0",
]

[project.urls]
Homepage = "https://github.com/sentinel/cybersecurity-triage"
Repository = "https://github.com/sentinel/cybersecurity-triage"
Documentation = "https://sentinel.readthedocs.io"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["sentinel"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "feedparser.*",
    "boto3.*",
    "botocore.*",
    "moto.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
]